Revert "Release v1.0.0"
