--- .github/workflows/lint.yaml
@@ -1,26 +1,0 @@
-name: Lint
-
-on:
-  push:
-    branches:
-      - main
-      - release
-  pull_request:
-
-permissions:
-  contents: read
-
-jobs:
-  lint:
-    name: Lint
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-go@v4
-        with:
-          go-version: "1.21"
-          cache: false
-      - name: Lint
-        uses: golangci/golangci-lint-action@v3
-        with:
-          version: v1.55
