+++ .github/workflows/release.yaml
@@ -11,25 +11,25 @@ jobs:
   release:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-go@v4
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Set up Go
+        uses: actions/setup-go@v4
         with:
-          go-version: 1.21
+          go-version: 1.21.x
       - name: Build with xgo
         uses: crazy-max/ghaction-xgo@v3
         with:
           xgo_version: latest
           go_version: 1.21.x
-          pkg: cmd/quickval/
+          pkg: cmd/quickval/quickval.go
           dest: build
           prefix: quickval
           targets: linux/amd64,darwin/amd64,darwin/arm64,windows/amd64,
           v: true
           ldflags: -s -w
-      - name: Use Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: "lts/*"
+      - name: Update Tickers
+        run: go run ./cmd/fetch_tickers/fetch_tickers.go --api-key=${{ secrets.QUICKFS_KEY }}
       - name: Install Dependencies
         run: npm install -g semantic-release @semantic-release/exec @semantic-release/git @semantic-release/github
       - name: Semantic Release
