--- .github/workflows/test.yaml
@@ -1,22 +1,0 @@
-name: Test
-
-on:
-  pull_request:
-    branches:
-      - main
-
-permissions:
-  contents: read
-
-jobs:
-  test:
-    name: Test
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - uses: actions/setup-go@v4
-        with:
-          go-version: "1.21"
-          cache: false
-      - name: Test
-        run: go test ./...
