+++ .github/workflows/update-tickers.yaml
@@ -3,19 +3,11 @@ name: Update Tickers
 on:
   schedule:
     - cron: "13 0 * * 1-5"
-  push:
-    branches:
-      - main
-    paths-ignore:
-      - "tickers/**"
 
 jobs:
   fetch:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/setup-go@v4
-        with:
-          go-version: 1.21
       - name: Checkout
         uses: actions/checkout@v4
         with:
