--- .gitignore
@@ -1,4 +1,0 @@
-.DS_Store
-
-./quickval
-
