+++ internal/output/output.go
@@ -42,6 +42,7 @@ func (w *Writer) Data(data *quickfs.Data) {
 	}
 
 	w.table.Append([]string{"", ""})
+
 }
 
 func (w *Writer) WACC(rate float64, erp float64, rfr float64, data *quickfs.Data) {
@@ -65,12 +66,11 @@ func (w *Writer) WACC(rate float64, erp float64, rfr float64, data *quickfs.Data
 	w.table.Append([]string{"", ""})
 }
 
-func (w *Writer) Projected(projected []int, growthRate float64, expectedReturn float64) {
+func (w *Writer) Projected(projected []int, growthRate float64) {
 	w.table.Append([]string{"", ""})
 	w.table.Append([]string{"PROJECTIONS", ""})
 	w.table.Append([]string{"----------------------------------------", "------------------"})
 	w.table.Append([]string{"Growth Rate", fmt.Sprintf("%.2f", growthRate)})
-	w.table.Append([]string{"Expected Return", fmt.Sprintf("%.2f", expectedReturn)})
 	w.table.Append([]string{"", ""})
 
 	// append projections
@@ -86,7 +86,7 @@ func (w *Writer) DiscountRate(rate float64) {
 	w.table.Append([]string{"", ""})
 	w.table.Append([]string{"----------------------------------------", "------------------"})
 	w.table.Append([]string{"DISCOUNT RATE", ""})
-	w.table.Append([]string{"User Specified", fmt.Sprintf("%.2f", rate)})
+	w.table.Append([]string{"Set by user", fmt.Sprintf("%.2f", rate)})
 
 	w.table.Append([]string{"", ""})
 }
