+++ tickers/AU.json
@@ -3235,6 +3235,5 @@
    "LU7:AU",
    "C79:AU",
    "1AI:AU",
-   "WMI:AU",
-   "EXT:AU"
+   "WMI:AU"
 ]
