+++ tickers/BE.json
@@ -114,9 +114,5 @@
    "ASCE:BE",
    "SHUR:BE",
    "BELYS:BE",
-   "BELU:BE",
-   "CLEX:BE",
-   "ALIMG:BE",
-   "CAND:BE",
-   "MOP:BE"
+   "BELU:BE"
 ]
