+++ tickers/CA.json
@@ -18,6 +18,7 @@
    "WG:CA",
    "NERD:CA",
    "XPHY:CA",
+   "RARE:CA",
    "FENX:CA",
    "NEON:CA",
    "MLCL:CA",
@@ -975,9 +976,7 @@
    "ESAU:CA",
    "LUX:CA",
    "PATH:CA",
-   "ER:CA",
    "LIVE:CA",
-   "AFF:CA",
    "AGC:CA",
    "M:CA",
    "HZ:CA",
@@ -991,71 +990,6 @@
    "RGX:CA",
    "TCEC:CA",
    "HMC:CA",
-   "AIG:CA",
-   "PXE:CA",
-   "LLM:CA",
-   "SASK:CA",
-   "GAND:CA",
-   "KCL:CA",
-   "CASC:CA",
-   "SPIR:CA",
-   "CYBT:CA",
-   "SPAI:CA",
-   "BNXT:CA",
-   "RI:CA",
-   "TUGA:CA",
-   "NBRK:CA",
-   "OBRN:CA",
-   "GRCM:CA",
-   "CISC:CA",
-   "SPOD:CA",
-   "METL:CA",
-   "VTAL:CA",
-   "HERB:CA",
-   "DCNN:CA",
-   "GEGC:CA",
-   "ENRT:CA",
-   "MLKM:CA",
-   "PRME:CA",
-   "AMB:CA",
-   "SUNN:CA",
-   "GDIG:CA",
-   "EBR:CA",
-   "RSH:CA",
-   "SCM:CA",
-   "WSR:CA",
-   "DEME:CA",
-   "VRTX:CA",
-   "HML:CA",
-   "WEBV:CA",
-   "WOGC:CA",
-   "SF:CA",
-   "CMT:CA",
-   "WMC:CA",
-   "MOOD:CA",
-   "MOLT:CA",
-   "ATCM:CA",
-   "LEEF:CA",
-   "AIC:CA",
-   "HTRC:CA",
-   "BG:CA",
-   "WISE:CA",
-   "LCR:CA",
-   "NF:CA",
-   "UGH:CA",
-   "NARA:CA",
-   "CACR.A:CA",
-   "HWTN:CA",
-   "SPTZ:CA",
-   "CRPC:CA",
-   "MESC:CA",
-   "ABRT:CA",
-   "MBIO:CA",
-   "MOSS:CA",
-   "PPB:CA",
-   "COMT:CA",
-   "LTHM:CA",
-   "BZAM:CA",
    "CCHW:CA",
    "STCK:CA",
    "PSK:CA",
@@ -1512,6 +1446,7 @@
    "ETG:CA",
    "ESI:CA",
    "ERF:CA",
+   "ER:CA",
    "EOX:CA",
    "ENGH:CA",
    "ENB:CA",
@@ -1973,6 +1908,7 @@
    "PJC.A:CA",
    "PRW:CA",
    "PSN:CA",
+   "PUR:CA",
    "PXX:CA",
    "QRM:CA",
    "R:CA",
@@ -2373,8 +2309,6 @@
    "TAL:CA",
    "SOIL:CA",
    "PRB:CA",
-   "HTL:CA",
-   "CNL:CA",
    "QIPT:CA",
    "LIRC:CA",
    "GATO:CA",
@@ -2383,35 +2317,8 @@
    "ECOR:CA",
    "GEC:CA",
    "HC:CA",
-   "LIQD:CA",
-   "IFSS:CA",
-   "SHWZ:CA",
-   "INXD:CA",
-   "VOXR:CA",
-   "NFT:CA",
-   "BCBN:CA",
-   "AAUC:CA",
-   "IBER:CA",
-   "BTQ:CA",
-   "AMEN:CA",
-   "EVTG:CA",
-   "ILLM:CA",
-   "UBER:CA",
-   "TSND:CA",
-   "VISA:CA",
-   "CRM:CA",
-   "ATRL:CA",
-   "CSCO:CA",
-   "DIS:CA",
-   "HD:CA",
-   "MATR:CA",
-   "NVDA:CA",
-   "PFE:CA",
-   "SBUX:CA",
-   "RARE:CA",
    "SGD:CA",
    "MLP:CA",
-   "PUR:CA",
    "SGQ:CA",
    "CHS:CA",
    "SGMD:CA",
@@ -2570,6 +2477,7 @@
    "AET:CA",
    "AFCC:CA",
    "PEAK:CA",
+   "AFF:CA",
    "AFM:CA",
    "AGB:CA",
    "AGG:CA",
@@ -4724,6 +4632,7 @@
    "CWM:CA",
    "CRE:CA",
    "ECR:CA",
+   "HTL:CA",
    "LMG:CA",
    "MOON:CA",
    "TCA:CA",
@@ -5002,6 +4911,7 @@
    "MKA:CA",
    "MSR:CA",
    "WWT:CA",
+   "CNL:CA",
    "KNC:CA",
    "JE:CA",
    "NAU:CA",
@@ -5334,48 +5244,5 @@
    "LCE:CA",
    "SNV.H:CA",
    "WRLG:CA",
-   "ADG.H:CA",
-   "AISX:CA",
-   "OSIX:CA",
-   "AVR:CA",
-   "MEEC:CA",
-   "COHO:CA",
-   "TWY.H:CA",
-   "PTFY:CA",
-   "GNFI:CA",
-   "ABA:CA",
-   "HMPG:CA",
-   "WOLF:CA",
-   "KIB:CA",
-   "GRD:CA",
-   "MTK:CA",
-   "CLIP:CA",
-   "INSP:CA",
-   "VMXX:CA",
-   "HPM:CA",
-   "EML:CA",
-   "NOAL:CA",
-   "CLIC:CA",
-   "INFM:CA",
-   "CELL:CA",
-   "ZLTO:CA",
-   "PNRL:CA",
-   "VOL:CA",
-   "RRL:CA",
-   "SPZ:CA",
-   "SLG:CA",
-   "JHC:CA",
-   "CGC:CA",
-   "NICN:CA",
-   "GHML:CA",
-   "CART.H:CA",
-   "RAGE:CA",
-   "EUP:CA",
-   "LUCA:CA",
-   "RUSH:CA",
-   "BONE:CA",
-   "BLOK.H:CA",
-   "KLDC:CA",
-   "TVC.H:CA",
-   "VLT:CA"
+   "ADG.H:CA"
 ]
