+++ tickers/CH.json
@@ -221,6 +221,5 @@
    "AMS:CH",
    "VT5:CH",
    "IGPH:CH",
-   "REHN:CH",
-   "SHLTN:CH"
+   "REHN:CH"
 ]
