+++ tickers/DE.json
@@ -4,9 +4,6 @@
    "ECF:DE",
    "EXJ:DE",
    "HBD1:DE",
-   "FPH:DE",
-   "GGS:DE",
-   "1HQ:DE",
    "TVD6:DE",
    "CPU2:DE",
    "CSQ:DE",
@@ -14,19 +11,20 @@
    "CPX:DE",
    "ALX:DE",
    "UCA1:DE",
+   "6I4:DE",
    "BMM:DE",
    "BSS:DE",
    "HYI:DE",
    "FEW:DE",
    "SIM0:DE",
    "5MP1:DE",
+   "VBX:DE",
    "PHH2:DE",
    "HG1:DE",
    "FVI:DE",
    "DCIK:DE",
    "MSH:DE",
    "SCM:DE",
-   "WWG:DE",
    "ADC:DE",
    "RTML:DE",
    "PEH:DE",
@@ -55,6 +53,7 @@
    "LEC:DE",
    "KGR:DE",
    "SSH:DE",
+   "NWX:DE",
    "NUQA:DE",
    "BBI:DE",
    "T2G:DE",
@@ -65,37 +64,7 @@
    "011:DE",
    "SHWK:DE",
    "TCU:DE",
-   "GTY:DE",
    "NLM:DE",
-   "2INV:DE",
-   "SGU:DE",
-   "82Y:DE",
-   "ABA:DE",
-   "4880:DE",
-   "PKB:DE",
-   "TSS:DE",
-   "MBH3:DE",
-   "VMR1:DE",
-   "U1DA:DE",
-   "TUF:DE",
-   "B9B:DE",
-   "PU11:DE",
-   "SZ50:DE",
-   "BFV:DE",
-   "C0M:DE",
-   "GAG:DE",
-   "SBE:DE",
-   "SMB:DE",
-   "GRF:DE",
-   "VSC:DE",
-   "Z29:DE",
-   "A62:DE",
-   "HVB:DE",
-   "H9Y:DE",
-   "AGR:DE",
-   "BLH:DE",
-   "9JK:DE",
-   "KTEK:DE",
    "FDD:DE",
    "TAV:DE",
    "CLS1:DE",
@@ -106,7 +75,6 @@
    "HHX:DE",
    "NEP:DE",
    "KBU:DE",
-   "VQT:DE",
    "ZPF:DE",
    "TLG:DE",
    "SCI:DE",
@@ -114,51 +82,25 @@
    "HBM:DE",
    "C9T:DE",
    "OEL:DE",
-   "NWX:DE",
    "OSR:DE",
    "I8CK:DE",
    "NST:DE",
-   "B5A:DE",
-   "TC1:DE",
-   "H24:DE",
-   "EBGK:DE",
-   "ABT:DE",
-   "TBX0:DE",
-   "VVX:DE",
-   "GWK3:DE",
-   "9AUA:DE",
-   "CAG0:DE",
    "RTV:DE",
    "HRU:DE",
    "VHO:DE",
    "DLB:DE",
+   "AAD:DE",
    "HMU:DE",
-   "LSX:DE",
-   "6I4:DE",
+   "1HQ:DE",
    "BMP:DE",
+   "WWG:DE",
    "MUT3:DE",
+   "BIO3:DE",
    "RCMN:DE",
    "FAK:DE",
    "0E3:DE",
+   "GTY:DE",
    "DWB:DE",
-   "GBQ:DE",
-   "TLIK:DE",
-   "QS6A:DE",
-   "EFS3:DE",
-   "SPB:DE",
-   "TIS:DE",
-   "STG:DE",
-   "GTK:DE",
-   "NXR:DE",
-   "ALXA:DE",
-   "ECX:DE",
-   "CTNK:DE",
-   "N2F0:DE",
-   "LNSX:DE",
-   "E2N:DE",
-   "EUX:DE",
-   "KSW:DE",
-   "ATW:DE",
    "BWB:DE",
    "VAR1:DE",
    "SGF:DE",
@@ -186,6 +128,7 @@
    "PCZ:DE",
    "AAG:DE",
    "KA8:DE",
+   "VQT:DE",
    "PDA:DE",
    "IBU:DE",
    "CEA:DE",
@@ -229,7 +172,6 @@
    "ADL:DE",
    "ACX:DE",
    "ACWN:DE",
-   "AAD:DE",
    "A7A:DE",
    "A6T:DE",
    "A4Y:DE",
@@ -285,6 +227,7 @@
    "FEV:DE",
    "FIE:DE",
    "FME:DE",
+   "FPH:DE",
    "FRA:DE",
    "FRE:DE",
    "FRS:DE",
@@ -293,6 +236,7 @@
    "G1A:DE",
    "GBF:DE",
    "GFT:DE",
+   "GGS:DE",
    "GKS:DE",
    "GLJ:DE",
    "GME:DE",
@@ -329,6 +273,7 @@
    "LEO:DE",
    "LHA:DE",
    "LPK:DE",
+   "LSX:DE",
    "LXS:DE",
    "M3BK:DE",
    "M4N:DE",
@@ -432,7 +377,6 @@
    "M8G:DE",
    "AT1:DE",
    "WIG1:DE",
-   "VBX:DE",
    "LRND:DE",
    "CORE:DE",
    "EMH1:DE",
@@ -449,7 +393,6 @@
    "SUR:DE",
    "RTC:DE",
    "PA8K:DE",
-   "BIO3:DE",
    "MRK:DE",
    "MED:DE",
    "KTN:DE",
@@ -470,135 +413,5 @@
    "APM:DE",
    "TGHN:DE",
    "O5G:DE",
-   "DHL:DE",
-   "ALG:DE",
-   "PBY:DE",
-   "O4B:DE",
-   "DTD2:DE",
-   "EVD:DE",
-   "TUI1:DE",
-   "SOW:DE",
-   "JUN3:DE",
-   "AMMN:DE",
-   "NIIN:DE",
-   "M5S:DE",
-   "3D6:DE",
-   "S9I:DE",
-   "NUVA:DE",
-   "KSC:DE",
-   "R1B:DE",
-   "LIK:DE",
-   "APPH:DE",
-   "123F:DE",
-   "SHL:DE",
-   "8TRA:DE",
-   "M0YN:DE",
-   "KTA:DE",
-   "S4AA:DE",
-   "NWO:DE",
-   "VNA:DE",
-   "BIKE:DE",
-   "27N0:DE",
-   "P911:DE",
-   "VH2:DE",
-   "CDZ0:DE",
-   "MRX:DE",
-   "HHHA:DE",
-   "PRY:DE",
-   "TIMA:DE",
-   "CEK:DE",
-   "2GI:DE",
-   "PO1:DE",
-   "TMV:DE",
-   "LDF:DE",
-   "PSG:DE",
-   "SBX:DE",
-   "HGEA:DE",
-   "NCH2:DE",
-   "AG1:DE",
-   "MA10:DE",
-   "AB9:DE",
-   "VRL:DE",
-   "NFN:DE",
-   "CYR:DE",
-   "1SXP:DE",
-   "MPSC:DE",
-   "WEW:DE",
-   "MUX:DE",
-   "JY0:DE",
-   "KBX:DE",
-   "RY8:DE",
-   "VTSC:DE",
-   "BIG1:DE",
-   "SYAB:DE",
-   "ACT:DE",
-   "DMP:DE",
-   "XTP:DE",
-   "KWG:DE",
-   "JZ6:DE",
-   "7V0:DE",
-   "ILM1:DE",
-   "PBB:DE",
-   "HP3A:DE",
-   "ENR:DE",
-   "HYQ:DE",
-   "INS:DE",
-   "FC9:DE",
-   "BOG:DE",
-   "14D:DE",
-   "DFV:DE",
-   "VEZ:DE",
-   "93M1:DE",
-   "BSL:DE",
-   "DTG:DE",
-   "NA9:DE",
-   "DWS:DE",
-   "HRPK:DE",
-   "TRU:DE",
-   "NF4:DE",
-   "C3RY:DE",
-   "CY1K:DE",
-   "COP:DE",
-   "BWQ:DE",
-   "EXL:DE",
-   "VIH1:DE",
-   "PUM:DE",
-   "DEQ:DE",
-   "SFQ:DE",
-   "HFG:DE",
-   "SF3:DE",
-   "ITN:DE",
-   "H2O:DE",
-   "HAG:DE",
-   "WSU:DE",
-   "DEF:DE",
-   "BKHT:DE",
-   "EBE:DE",
-   "PAG:DE",
-   "1COV:DE",
-   "MBB:DE",
-   "ADE:DE",
-   "TTO:DE",
-   "UBK:DE",
-   "CHG:DE",
-   "TPG:DE",
-   "UN0:DE",
-   "EAD:DE",
-   "TR9:DE",
-   "D6H:DE",
-   "BRNK:DE",
-   "HXCK:DE",
-   "YOU:DE",
-   "WCH:DE",
-   "AMI:DE",
-   "UDC:DE",
-   "SJJ:DE",
-   "N4G:DE",
-   "G24:DE",
-   "ABO:DE",
-   "ETE:DE",
-   "EXC:DE",
-   "NVM:DE",
-   "BFSA:DE",
-   "GFG:DE"
+   "DHL:DE"
 ]
