+++ tickers/DK.json
@@ -171,7 +171,5 @@
    "DATA:DK",
    "HRC:DK",
    "GYLD.A:DK",
-   "AGF.B:DK",
-   "GYLD.B:DK",
-   "CARL.B:DK"
+   "AGF.B:DK"
 ]
