+++ tickers/ES.json
@@ -130,12 +130,5 @@
    "NYE:ES",
    "EDR:ES",
    "VANA:ES",
-   "LAB:ES",
-   "GGR:ES",
-   "NAT:ES",
-   "OPDE:ES",
-   "ALC:ES",
-   "ETC:ES",
-   "INDXA:ES",
-   "ALQ:ES"
+   "LAB:ES"
 ]
