+++ tickers/FR.json
@@ -669,23 +669,5 @@
    "MLVSY:FR",
    "URW:FR",
    "MLCFD:FR",
-   "CIV:FR",
-   "ALSOG:FR",
-   "ALMKT:FR",
-   "ALMEX:FR",
-   "MLIFS:FR",
-   "ALBOU:FR",
-   "PHXM:FR",
-   "MHM:FR",
-   "ALEXP:FR",
-   "ALTD:FR",
-   "MLHPE:FR",
-   "MLLCB:FR",
-   "MLVST:FR",
-   "MLCHE:FR",
-   "ALSRS:FR",
-   "ALPM:FR",
-   "MLLOI:FR",
-   "MLCAC:FR",
-   "SESG:FR"
+   "CIV:FR"
 ]
