+++ tickers/GR.json
@@ -83,11 +83,5 @@
    "DOPPLER:GR",
    "AAAK:GR",
    "DIMAND:GR",
-   "ACAG:GR",
-   "DAIOS:GR",
-   "INLIF:GR",
-   "TRESTATES:GR",
-   "OPTIMA:GR",
-   "ORILINA:GR",
-   "MERKO:GR"
+   "ACAG:GR"
 ]
