+++ tickers/IT.json
@@ -374,6 +374,7 @@
    "NUS:IT",
    "MTS:IT",
    "IMP:IT",
+   "1.TIME:IT",
    "ICALL:IT",
    "RWY:IT",
    "OMER:IT",
@@ -414,43 +415,5 @@
    "PAG:IT",
    "LASIA:IT",
    "IVN:IT",
-   "PMT:IT",
-   "ERFO:IT",
-   "GSR:IT",
-   "TMP:IT",
-   "EVA:IT",
-   "FIRE:IT",
-   "VLT:IT",
-   "HQF:IT",
-   "YOLO:IT",
-   "ELG:IT",
-   "SAC:IT",
-   "ART:IT",
-   "EPRB:IT",
-   "NVP:IT",
-   "GEL:IT",
-   "TCM:IT",
-   "FDA:IT",
-   "OLI:IT",
-   "GML:IT",
-   "FAE:IT",
-   "MEV:IT",
-   "GM:IT",
-   "DBA:IT",
-   "RDUE:IT",
-   "EAV:IT",
-   "EPR:IT",
-   "SIAV:IT",
-   "VARV:IT",
-   "XHS:IT",
-   "ELSA:IT",
-   "YACHT:IT",
-   "POZ:IT",
-   "TIME:IT",
-   "FCH:IT",
-   "VLC:IT",
-   "RKT:IT",
-   "TWEP:IT",
-   "GBUS:IT",
-   "CORE:IT"
+   "PMT:IT"
 ]
