+++ tickers/PL.json
@@ -720,11 +720,5 @@
    "END:PL",
    "BLF:PL",
    "RES:PL",
-   "CBF:PL",
-   "GAR:PL",
-   "GRE:PL",
-   "PLT:PL",
-   "ATS:PL",
-   "UFC:PL",
-   "GOV:PL"
+   "CBF:PL"
 ]
