+++ tickers/SE.json
@@ -428,6 +428,7 @@
    "SAS:SE",
    "SAVE:SE",
    "SBOK:SE",
+   "SENS:SE",
    "SEZI:SE",
    "SINT:SE",
    "SIVE:SE",
@@ -990,23 +991,5 @@
    "REFINE:SE",
    "BUFAB:SE",
    "NOHARM:SE",
-   "WPTG.B:SE",
-   "BONZUN:SE",
-   "TOL:SE",
-   "IMSYS:SE",
-   "BIOS:SE",
-   "EBLITZ:SE",
-   "ALONNR.B:SE",
-   "EDYOU:SE",
-   "YUBICO:SE",
-   "PIXEL:SE",
-   "HMPLY:SE",
-   "CHEF:SE",
-   "RUSTA:SE",
-   "SIBEK:SE",
-   "OODA:SE",
-   "BIOVIT:SE",
-   "COPP:SE",
-   "NANOC.B:SE",
-   "SAFETY.B:SE"
+   "WPTG.B:SE"
 ]
